Put eth/beacon/enums/ objects in one file
