--- eth/beacon/enums/signature_domain.py
@@ -1,8 +1,0 @@
-from enum import IntEnum
-
-
-class SignatureDomain(IntEnum):
-    DOMAIN_DEPOSIT = 0
-    DOMAIN_ATTESTATION = 1
-    DOMAIN_PROPOSAL = 2
-    DOMAIN_LOGOUT = 3
