--- eth/beacon/enums/special_record_types.py
@@ -1,8 +1,0 @@
-from enum import IntEnum
-
-
-class SpecialRecordTypes(IntEnum):
-    VOLUNTARY_EXIT = 0
-    CASPER_SLASHING = 1
-    PROPOSER_SLASHING = 2
-    DEPOSIT_PROOF = 3
