--- eth/beacon/enums/validator_registry_delta_flags.py
@@ -1,6 +1,0 @@
-from enum import IntEnum
-
-
-class ValidatorRegistryDeltaFlag(IntEnum):
-    ACTIVATION = 0
-    EXIT = 1
