+++ tests/beacon/helpers.py
@@ -1,4 +1,4 @@
-from eth.beacon.enums.validator_status_codes import (
+from eth.beacon.enums import (
     ValidatorStatusCode,
 )
 from eth.beacon.types.validator_records import (
